Merge conflicts resolved.

+++ server/api/products.js
@@ -1,5 +1,6 @@
 const router = require('express').Router()
 const {Product} = require('../db/models')
+const isAdmin = require('../../helpers/isAdminHelperFunc')
 module.exports = router
 
 // GET /api/products
@@ -27,8 +28,14 @@ router.get('/:productId', async (req, res, next) => {
 // POST /api/products
 router.post('/', async (req, res, next) => {
   try {
-    const newProduct = await Product.create(req.body)
-    res.json(newProduct)
+    if (req.user === undefined) {
+      res.sendStatus(404)
+    } else if (isAdmin(req.user.id)) {
+      const newProduct = await Product.create(req.body)
+      res.json(newProduct)
+    } else {
+      res.sendStatus(404)
+    }
   } catch (error) {
     next(error)
   }
@@ -37,10 +44,15 @@ router.post('/', async (req, res, next) => {
 // PUT /api/products/:productId
 router.put('/:productId', async (req, res, next) => {
   try {
-    console.log('router: productId = ', productId)
-    const product = await Product.findByPk(req.params.productId)
-    const updatedProduct = await product.update(req.body)
-    res.status(200).send(updatedProduct)
+    if (req.user === undefined) {
+      res.sendStatus(404)
+    } else if (isAdmin(req.user.id)) {
+      const product = await Product.findByPk(req.params.productId)
+      const updatedProduct = await product.update(req.body)
+      res.status(200).send(updatedProduct)
+    } else {
+      res.sendStatus(404)
+    }
   } catch (error) {
     next(error)
   }
@@ -49,9 +61,15 @@ router.put('/:productId', async (req, res, next) => {
 // DELETE /api/products/:productId
 router.delete('/:productId', async (req, res, next) => {
   try {
-    const id = req.params.productId
-    await Product.destroy({where: {id}})
-    res.sendStatus(204)
+    if (req.user === undefined) {
+      res.sendStatus(404)
+    } else if (isAdmin(req.user.id)) {
+      const id = req.params.productId
+      await Product.destroy({where: {id}})
+      res.sendStatus(204)
+    } else {
+      res.sendStatus(404)
+    }
   } catch (error) {
     next(error)
   }
